29/05/2023

+++ backend/controller/reportProblem.go
@@ -365,17 +365,12 @@ func UpdateReportProblem(c *gin.Context) {
 		}
 		reportProblem.FileUpload = fileUpload
 	}
-	if reportProblem.FileUploadID == nil {
-		// ตรวจสอบว่ามีไฟล์ที่เกี่ยวข้องอยู่หรือไม่
-		if reportProblem.FileUpload.ID != 0 {
-			// ลบไฟล์ออกจากฐานข้อมูล
-			if err := entity.DB().Delete(&reportProblem.FileUpload).Error; err != nil {
-				c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-				return
-			}
+	if reportProblem.FileUploadID == nil && reportProblem.FileUpload.ID != 0 {
+		// Delete old file upload from the database
+		if err := entity.DB().Delete(&reportProblem.FileUpload).Error; err != nil {
+			c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+			return
 		}
-		reportProblem.FileUploadID = nil
-		reportProblem.FileUpload = entity.FileUpload{}
 	}
 
 	update := entity.ReportProblem{
